three_point based indices
